refactor: rename basic_java_types to java_basic_types
Rename all occurrences of basic_java_types to java_basic_types for consistency and clarity across the codebase. This improves naming alignment with other modules and enhances code readability.
