+++ src/commands/mod.rs
@@ -26,9 +26,9 @@ use crate::{
     package_name_validator::validate_package_name,
   },
   common::types::{
-    basic_field_config::BasicFieldConfig, basic_java_types::JavaBasicType,
-    cascade_type::CascadeType, collection_type::CollectionType, enum_field_config::EnumFieldConfig,
-    fetch_type::FetchType, id_field_config::IdFieldConfig, java_enum_type::JavaEnumType,
+    basic_field_config::BasicFieldConfig, cascade_type::CascadeType,
+    collection_type::CollectionType, enum_field_config::EnumFieldConfig, fetch_type::FetchType,
+    id_field_config::IdFieldConfig, java_basic_types::JavaBasicType, java_enum_type::JavaEnumType,
     java_field_temporal::JavaFieldTemporal, java_field_time_zone_storage::JavaFieldTimeZoneStorage,
     java_file_type::JavaFileType, java_id_generation::JavaIdGeneration,
     java_id_generation_type::JavaIdGenerationType,
