+++ src/commands/services/create_jpa_one_to_one_relationship_service.rs
@@ -10,8 +10,8 @@ use crate::common::ts_file::TSFile;
 use crate::common::types::annotation_config::AnnotationConfig;
 use crate::common::types::cascade_type::CascadeType;
 use crate::common::types::entity_side::EntitySide;
-use crate::common::types::field_types::FieldInsertionPosition;
 use crate::common::types::import_types::ImportInsertionPosition;
+use crate::common::types::java_basic_types::FieldInsertionPosition;
 use crate::common::types::java_source_directory_type::JavaSourceDirectoryType;
 use crate::common::types::java_visibility_modifier::JavaVisibilityModifier;
 use crate::common::types::mapping_type::MappingType;
